sidebar complete 99
