+++ src/components/ProfilesLi.jsx
@@ -1,8 +1,22 @@
+import { useDispatch } from "react-redux";
 import { useNavigate } from "react-router-dom";
+import { ADD_CLICKED_PROFILE, fetchProfile } from "../redux/actions/actions";
 import GreyBorderBtn from "./GreyBorderBtn";
 
 const ProfilesLi = (props) => {
   const navigate = useNavigate();
+  const action = ADD_CLICKED_PROFILE;
+  const dispatch = useDispatch();
+  const id = props.profile._id;
+  const endPoint = "https://striveschool-api.herokuapp.com/api/profile/";
+  const accessToken =
+    "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2Mzk4NGEyNzQwNWJkYTAwMTUwOTE4NDQiLCJpYXQiOjE2NzA5MjQ4MzksImV4cCI6MTY3MjEzNDQzOX0.x2Rft_8jW0eH4mFzHLq669IFCzGAFGCn7LuvHCf2udU";
+  const options = {
+    method: "GET",
+    headers: {
+      Authorization: "Bearer " + accessToken,
+    },
+  };
 
   return (
     <li className="sidebarLi d-flex">
@@ -12,7 +26,10 @@ const ProfilesLi = (props) => {
             className="xs-profiles pointer"
             src={props.profile.image}
             alt="mini-profile-pics"
-            onClick={() => navigate(`/profile/${props.profile._id}`)}
+            onClick={() => {
+              navigate(`/profile/${props.profile._id}`);
+              dispatch(fetchProfile(endPoint, options, id, action));
+            }}
           />
         </div>
       </div>
@@ -21,7 +38,10 @@ const ProfilesLi = (props) => {
         <li className="pt-3">
           <span
             className="fw-bold pointer"
-            onClick={() => navigate(`/profile/${props.profile._id}`)}
+            onClick={() => {
+              navigate(`/profile/${props.profile._id}`);
+              dispatch(fetchProfile(endPoint, options, id, action));
+            }}
           >
             {props.profile.name} {props.profile.surname} |{" "}
           </span>
