+++ src/redux/reducers/profilesReducer.jsx
@@ -1,7 +1,8 @@
-import { ADD_ALL_PROFILES } from "../actions/actions";
+import { ADD_ALL_PROFILES, ADD_CLICKED_PROFILE } from "../actions/actions";
 
 const initialState = {
   allProfiles: [],
+  clickedProfile: [],
 };
 
 const profilesReducer = (state = initialState, action) => {
@@ -10,6 +11,13 @@ const profilesReducer = (state = initialState, action) => {
       return {
         ...state,
         allProfiles: action.payload,
+      };
+    }
+
+    case ADD_CLICKED_PROFILE: {
+      return {
+        ...state,
+        clickedProfile: action.payload,
       };
     }
